+++ kilometer-backend-domain/src/main/java/com/kilometer/domain/item/dto/ItemUpdateRequest.java
@@ -1,12 +1,11 @@
 package com.kilometer.domain.item.dto;
 
-import com.kilometer.domain.item.ExhibitionType;
-import com.kilometer.domain.item.ExposureType;
-import com.kilometer.domain.item.FeeType;
-import com.kilometer.domain.item.ItemDetail;
-import com.kilometer.domain.item.ItemDetailImage;
-import com.kilometer.domain.item.ItemEntity;
-import com.kilometer.domain.item.RegionType;
+import com.kilometer.domain.item.enumType.ExhibitionType;
+import com.kilometer.domain.item.enumType.ExposureType;
+import com.kilometer.domain.item.enumType.FeeType;
+import com.kilometer.domain.item.enumType.RegionType;
+import com.kilometer.domain.item.itemDetail.ItemDetail;
+import com.kilometer.domain.item.itemDetailImage.ItemDetailImage;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -51,39 +50,19 @@ public class ItemUpdateRequest {
     @Builder.Default
     private List<String> detailImageUrls = new ArrayList<>();
 
+    @Builder.Default
     private List<Long> deleteDetailImages = new ArrayList<>();
 
     public ItemDetail makeUpdateItemDetail() {
         return ItemDetail.builder()
-                .introduce(this.introduce)
-                .build();
+            .introduce(this.introduce)
+            .build();
     }
 
     public List<ItemDetailImage> makeUpdateItemDetailImage() {
         return this.detailImageUrls.stream()
-                .filter(StringUtils::hasText)
-                .map(ItemDetailImage::makeEntity)
-                .collect(Collectors.toList());
-    }
-
-    public ItemEntity makeUpdateItemEntity() {
-        return ItemEntity.builder()
-                .exhibitionType(this.getExhibitionType())
-                .exposureType(this.getExposureType())
-                .listImageUrl(this.getListImageUrl())
-                .thumbnailImageUrl(this.getThumbnailImageUrl())
-                .title(this.getTitle())
-                .startDate(this.getStartDate())
-                .endDate(this.getEndDate())
-                .placeName(this.getPlaceName())
-                .latitude(this.getLatitude())
-                .longitude(this.getLongitude())
-                .regionType(this.getRegionType())
-                .feeType(this.getFeeType())
-                .price(this.getPrice())
-                .homepageUrl(this.getHomepageUrl())
-                .operatingTime(this.getOperatingTime())
-                .ticketUrl(this.getTicketUrl())
-                .build();
+            .filter(StringUtils::hasText)
+            .map(ItemDetailImage::makeEntity)
+            .collect(Collectors.toList());
     }
 }
