+++ kilometer-backend-domain/src/main/java/com/kilometer/domain/item/enumType/ExposureType.java
@@ -1,13 +1,13 @@
-package com.kilometer.domain.item;
+package com.kilometer.domain.item.enumType;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum FeeType {
+public enum ExposureType {
 
-    FREE("무료"), COST("유료");
+    ON("전시"), OFF("미전시");
 
     private final String description;
 }
