+++ kilometer-backend-domain/src/main/java/com/kilometer/domain/item/enumType/RegionType.java
@@ -1,4 +1,4 @@
-package com.kilometer.domain.item;
+package com.kilometer.domain.item.enumType;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
