+++ kilometer-backend-domain/src/main/java/com/kilometer/domain/item/heart/ItemHeart.java
@@ -1,4 +1,4 @@
-package com.kilometer.domain.search.heart;
+package com.kilometer.domain.item.heart;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 @Builder
 @AllArgsConstructor
-public class ListItemHeart {
+public class ItemHeart {
     private boolean heartClicked;
     private String link;
     private long id;
