+++ kilometer-backend-domain/src/main/java/com/kilometer/domain/item/itemDetail/ItemDetailRepository.java
@@ -1,4 +1,4 @@
-package com.kilometer.domain.item;
+package com.kilometer.domain.item.itemDetail;
 
 import org.springframework.data.jpa.repository.JpaRepository;
